Account for possible non-weakrefs to receivers
By default django signals refer to receivers via weakrefs.  However, in the event that there is not a global reference to the receiver (e.g. when the receiver is an instance of a class) the receiver will be garbage collected.  This means we should use the `weak=False` flag.

However, the `show_events` assumes that all events are weakly referenced and this results in errors when they are not.